Removed Text Mesh Pro and got some camera movement working
Text mesh pro created a fuck load of errors as soon as it was imported so I removed it bc cba to fix right now.

Got the camera edge scrolling working and the W and S but A and D wont work? Not sure why, if someone could have a look I'd appriciate it